Refactor reservations feature.
