+++ src2/utilities/renderPopup.js
@@ -1,13 +1,15 @@
 //import '../reservationPopupStyles.css';
 import generatePopupDOM from './popupDOM.js';
-import listener from './helperFunctions/listener.js';
-import addNewReservation from './APIs/addNewReservation.js';
-import getAllReservations from './APIs/getAllReservations.js';
+import listener from './listener.js';
+import {addNewReservation} from './postData.js';
+import {getAllReservations} from './fetchData.js';
 import { renderReservations } from './renderlist.js';
-import count from '../../counterFunctions/reservationCounter.js';
+import count from '../counterFunctions/reservationCounter.js';
 
 let counter = 0;
 let reservations = [];
+const url = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/kUJtIKt0WlDGnehZIL7s/reservations';
+const getReservationsURL = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/kUJtIKt0WlDGnehZIL7s/reservations?item_id=';
 
 const reservationPopup = (obj) => {
   const Popup = document.createElement('div');
@@ -18,17 +20,17 @@ const reservationPopup = (obj) => {
   const dateStart = Popup.querySelector('.startdate-input');
   const dateEnd = Popup.querySelector('.enddate-input');
 
-  getAllReservations(obj.id)
+  getAllReservations(obj.id, getReservationsURL)
     .then((res) => res.json())
     .then((data) => {
       reservations = data;
       counter = count(data) || 0;
       renderReservations(reservations, counter);
-    });
+    }).catch(error => console.log(error));
 
   listener(closeBtn, 'click', () => Popup.remove());
   listener(reserve, 'click', () => {
-    addNewReservation(obj.id, userName.value, dateStart.value, dateEnd.value);
+    addNewReservation(obj.id, userName.value, dateStart.value, dateEnd.value, url);
   });
 
   return Popup;
